hub: move all modules from errata app to scan
Resolves: https://github.com/openscanhub/openscanhub/issues/199
