+++ osh/hub/osh_xmlrpc/worker.py
@@ -11,11 +11,11 @@ from kobo.django.upload.models import FileUpload
 from kobo.hub.decorators import validate_worker
 from kobo.hub.models import Task
 
-from osh.hub.errata.scanner import (BaseNotValidException, obtain_base,
-                                    prepare_base_scan)
 from osh.hub.scan.models import (SCAN_STATES, AnalyzerVersion, AppSettings,
                                  Profile, Scan, ScanBinding)
 from osh.hub.scan.notify import send_task_notification
+from osh.hub.scan.scanner import (BaseNotValidException, obtain_base,
+                                  prepare_base_scan)
 from osh.hub.scan.xmlrpc_helper import cancel_scan
 from osh.hub.scan.xmlrpc_helper import fail_scan as h_fail_scan
 from osh.hub.scan.xmlrpc_helper import finish_scan as h_finish_scan
